// src/types/mock-data.type.ts
export type MockData = {
  names: string[];
  descriptions: string[];
  previewPaths: string[];
  videoPaths: string[];
  actors: string[];
  directors: string[];
  users: string[];
  emails: string[];
  avatarPaths: string[];
  posterPaths: string[];
  backgroundImagePaths: string[];
  backgroundColors: string[];
};